% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaWold}
\alias{armaWold}
\title{Calculating Wold coefficient from ARMA model parameters}
\usage{
armaWold(n, a, b)
}
\arguments{
\item{n}{number}

\item{a}{AR model coefficient}

\item{b}{MA model coefficient}
}
\value{
wold coefficient
}
\description{
Calculating Wold coefficient from ARMA model parameters
}
\examples{
\dontrun{
std::vector<double> a = {0.5, 0.3, 0.2};
std::vector<double> b = {0.1, 0.2};
int n = 10;
armaWold(n, a, b)
}
}
