% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204133.R
\name{example}
\alias{example}
\title{An example of MA theoretical spectral density using Rcpp}
\description{
Use R package \code{stats} to generate MA model random number and apply to R functions (\code{ma_true_spectrum}).
}
\examples{
\dontrun{
set.seed(101)
xma <- arima.sim(model=list(ma=c(0.5, -0.4)), n = 100, sd = sqrt(4) )
ma_true_spectrum(NumericVector xma, int ngrid = 256, double sigma = 1.0)
}
}
