% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204133.R
\name{ma.true.spectrum}
\alias{ma.true.spectrum}
\title{MA theoretical spectral density using R}
\usage{
ma.true.spectrum(
  a,
  ngrid = 256,
  sigma = 1,
  tit = "True MA Spectral Density",
  plot.it = TRUE
)
}
\arguments{
\item{a}{MA matrix}

\item{ngrid}{Interval number}

\item{sigma}{error term}

\item{tit}{title of picture}

\item{plot.it}{whether plot}
}
\value{
frequencies spectrum ma_coefficients sigma
}
\description{
MA theoretical spectral density using R
}
\examples{
\dontrun{
set.seed(101)
 xma <- arima.sim(model=list(ma=c(0.5, -0.4)), n = 100, sd = sqrt(4) )
ma_true_spectrum(NumericVector xma, int ngrid = 256, double sigma = 1.0)
}
}
