\name{SA24204133-package}
\alias{SA24204133-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  R and Rcpp versions of gibbs sampler and vaac function are provided in this package (\code{gibbsR} and \code{vaccR} versus \code{gibbsC} and \code{vaccC})
}
\author{
Wenrui Hu

Maintainer: Wenrui Hu <1794773782@qq.com>
}
\keyword{ package }
